Initial pass at pulling events list from Meetup and displaying, addressing #4.
Droplet.client makes the request to the Meetup API, which returns a list in
JSON. The JSON is unpacked into instances of `MeetupEvent`. `MeetupEvent`
currently stores just the fields known to be needed for the template, and
implements `MustacheBoxable` so that it can be included directly in the
rendering context.

Created Mustache template and stylesheet for the page, with two filters to
make numerical data look good.

Added link to new page to homepage nav.

Added basic testing for new functionality: Mustache filters and `MeetupEvent`.

Needs further structural refinement: currently too much is done inside the
route handler.