Finishes #4 "Pull data from Meetup API"
The "/events" route handler now uses a MeetupAPI object, which makes the
connection to api.meetup.com. The MeetupAPI.getUpcomingEvents() method pulls
down future events.

The events are provided by Meetup.com as a JSON array. The JSON objects are
unpacked into a parallel array of MeetupEvent structs, which is passed back
out to the route handler to be used as the context for the view.

A private Configuration struct supplies MeetupAPI with the appropriate URL
information. It gets the URL pieces from the Config/meetup.json file,
accessing that file via the active Droplet's Config object: the Droplet
aleady knows where the config files are, rather than us having to find them
again. The Droplet's Config is passed in to MeetupAPI on its creation, which
hands it off to its Configuration.